ndp: Move PacketNdp out of icmp
