+++ src/CMakeLists.txt
@@ -48,7 +48,7 @@ set(OS_OBJECTS
   net/tcp/write_queue.cpp net/tcp/rttm.cpp net/tcp/listener.cpp
   net/tcp/read_buffer.cpp net/tcp/read_request.cpp net/tcp/stream.cpp
   net/ip4/icmp4.cpp net/ip4/udp.cpp net/ip4/udp_socket.cpp
-  net/ip6/ip6.cpp net/ip6/icmp6.cpp net/ip6/ndp.cpp net/ip6/slaac.cpp
+  net/ip6/ip6.cpp net/ip6/icmp6.cpp net/ip6/ndp.cpp net/ip6/packet_ndp.cpp net/ip6/slaac.cpp
   net/dns/dns.cpp net/dns/client.cpp net/dhcp/dh4client.cpp net/dhcp/dhcpd.cpp
   net/buffer_store.cpp net/inet.cpp
   net/super_stack.cpp net/configure.cpp net/conntrack.cpp net/vlan_manager.cpp
